Code format and refactor Quad.

+++ Engine/Graphics/OpenGL/GLPrimitive.c
@@ -1,11 +1,16 @@
 /*
- * Copyright (c) 2012-2018 scott.cgi All Rights Reserved.
+ * Copyright (c) 2012-2019 scott.cgi All Rights Reserved.
  *
- * This code is licensed under the MIT License.
+ * This code and its project Mojoc are licensed under [the MIT License],
+ * and the project Mojoc is a game engine hosted on github at [https://github.com/scottcgi/Mojoc],
+ * and the author's personal website is [https://scottcgi.github.io],
+ * and the author's email is [scott.cgi@qq.com].
  *
  * Since : 2014-2-26
+ * Update: 2019-1-24
  * Author: scott.cgi
  */
+
 
 #include "Engine/Graphics/OpenGL/GLPrimitive.h"
 #include "Engine/Graphics/OpenGL/Shader/ShaderPrimitive.h"
@@ -33,24 +38,24 @@ static inline void SetMatrix()
 static void DrawPoints(Array(float)* pointArr)
 {
     SetMatrix();
-    glVertexAttribPointer(AShaderPrimitive->attribPosition, 2, GL_FLOAT, false, 0, pointArr->data);
-    glDrawArrays(GL_POINTS, 0, pointArr->length >> 1);
+    glVertexAttribPointer((GLuint) AShaderPrimitive->attribPosition, 2, GL_FLOAT, false, 0, pointArr->data);
+    glDrawArrays(GL_POINTS, 0, pointArr->length >> 1); // NOLINT(hicpp-signed-bitwise)
 }
 
 
 static void DrawPolygon(Array(float)* vertexArr)
 {
     SetMatrix();
-    glVertexAttribPointer(AShaderPrimitive->attribPosition, 2, GL_FLOAT, false, 0, vertexArr->data);
-    glDrawArrays(GL_LINE_LOOP, 0, vertexArr->length >> 1);
+    glVertexAttribPointer((GLuint) AShaderPrimitive->attribPosition, 2, GL_FLOAT, false, 0, vertexArr->data);
+    glDrawArrays(GL_LINE_LOOP, 0, vertexArr->length >> 1); // NOLINT(hicpp-signed-bitwise)
 }
 
 
 static void DrawLines(Array(float)* vertexArr)
 {
     SetMatrix();
-    glVertexAttribPointer(AShaderPrimitive->attribPosition, 2, GL_FLOAT, false, 0, vertexArr->data);
-    glDrawArrays(GL_LINES, 0, vertexArr->length >> 1);
+    glVertexAttribPointer((GLuint) AShaderPrimitive->attribPosition, 2, GL_FLOAT, false, 0, vertexArr->data);
+    glDrawArrays(GL_LINES, 0, vertexArr->length >> 1); // NOLINT(hicpp-signed-bitwise)
 }
 
 
@@ -60,7 +65,7 @@ static void DrawRect(Rect* rect)
 
     glVertexAttribPointer
     (
-        AShaderPrimitive->attribPosition,
+        (GLuint) AShaderPrimitive->attribPosition,
         2,
         GL_FLOAT,
         false,
