+++ Engine/Graphics/OpenGL/Shader/ShaderPrimitive.h
@@ -1,14 +1,19 @@
 /*
- * Copyright (c) 2012-2018 scott.cgi All Rights Reserved.
+ * Copyright (c) 2012-2019 scott.cgi All Rights Reserved.
  *
- * This code is licensed under the MIT License.
+ * This code and its project Mojoc are licensed under [the MIT License],
+ * and the project Mojoc is a game engine hosted on github at [https://github.com/scottcgi/Mojoc],
+ * and the author's personal website is [https://scottcgi.github.io],
+ * and the author's email is [scott.cgi@qq.com].
  *
  * Since : 2014-2-26
+ * Update: 2019-1-24
  * Author: scott.cgi
  */
 
+
 #ifndef SHADER_PRIMITIVE_H
-#define SHADER_PRIMITIVE_H
+#define  SHADER_PRIMITIVE_H
 
 
 #include "Engine/Graphics/OpenGL/Platform/gl3.h"
@@ -17,6 +22,9 @@
 #include "Engine/Graphics/Draw/Color.h"
 
 
+/**
+ * Shader control for GLPrimitive.
+ */
 struct AShaderPrimitive
 {
     GLuint program;
@@ -25,7 +33,7 @@ struct AShaderPrimitive
     GLint  uniformColor;
     GLint  uniformPointSize;
 
-    void (*Use) (Matrix4* mvpMatrix, Color* color, GLint pointSize);
+    void (*Use) (Matrix4* mvpMatrix, Color* color, GLfloat pointSize);
     void (*Init)();
 };
